refactor: pair_history should only be available in webserver mode
